% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboGSE.R
\name{abbr}
\alias{abbr}
\title{Abbreviate GO term description}
\usage{
abbr(x)
}
\arguments{
\item{x}{A string describing a GO term}
}
\value{
The abbreviation of x
}
\description{
This function produces abbreviations in GO term description.
}
\keyword{internal}
