% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReactFixed.R
\name{addReactFixed}
\alias{addReactFixed}
\title{A fixed version of sybil::addReact}
\usage{
addReactFixed(
  model,
  id,
  met,
  Scoef,
  reversible = FALSE,
  lb = 0,
  ub = SYBIL_SETTINGS("MAXIMUM"),
  obj = 0,
  subSystem = NA,
  gprAssoc = NA,
  reactName = NA,
  metName = NA,
  metComp = NA
)
}
\arguments{
\item{model}{See \code{\link{addReact}}.}

\item{id}{See \code{\link{addReact}}.}

\item{met}{See \code{\link{addReact}}.}

\item{Scoef}{See \code{\link{addReact}}.}

\item{reversible}{See \code{\link{addReact}}.}

\item{lb}{See \code{\link{addReact}}.}

\item{ub}{See \code{\link{addReact}}.}

\item{obj}{See \code{\link{addReact}}.}

\item{subSystem}{See \code{\link{addReact}}.}

\item{gprAssoc}{See \code{\link{addReact}}.}

\item{reactName}{See \code{\link{addReact}}.}

\item{metName}{See \code{\link{addReact}}.}

\item{metComp}{See \code{\link{addReact}}.}
}
\value{
An object of class \code{modelorg}.
}
\description{
The function fixes sybil::addReact to avoid deprecated functions rBind and cBind.
}
\keyword{internal}
