% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprMaguire}
\alias{exprMaguire}
\title{Gene expression from Maguire et al. RNA-seq data}
\usage{
data(exprMaguire)
}
\description{
Gene expression from Maguire et al. RNA-seq data
}
\references{
Maguire, S. L. et al. (2014) Zinc finger transcription factors displaced SREBP proteins as the major Sterol regulators during Saccharomycotina evolution. PLoS Genet. 10, e1004076.
}
\author{
Maguire et al.
}
