% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescue.R
\name{generic}
\alias{generic}
\title{Relax model external source}
\usage{
generic(model)
}
\arguments{
\item{model}{An object of class \code{modelorg}.}
}
\value{
The \emph{generic} model.
}
\description{
This function returns the \emph{generic} form of a given model, where external reactions are rendered bidirectional and unrestricted.
}
\keyword{internal}
