% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iMK735}
\alias{iMK735}
\title{iMK735 metabolic models in hypoxia and normoxia}
\usage{
data(iMK735)
}
\description{
iMK735 metabolic models in hypoxia and normoxia
}
\references{
Kavšček, M., Bhutada, G., Madl, T. & Natter, K. (2015) Optimization of lipid production with a genome-scale model of Yarrowia lipolytica. BMC Syst. Biol. 9, 72.
}
\author{
Kavscek et al.
}
