% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboGSE.R
\name{maxArea}
\alias{maxArea}
\title{Compute maxium area between curves}
\usage{
maxArea(y, x = NULL)
}
\arguments{
\item{y}{A numeric matrix for all curves with each curve per column.}

\item{x}{A vector of increasing x-coordinates. Default, an equidistant vector of length \code{nrow(y)} from 0 to 1.}
}
\value{
The maximum area
}
\description{
This function computes the maximum area between multiple curves.
}
\keyword{internal}
