% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboGSE.R
\name{maxDist}
\alias{maxDist}
\title{Compute maximum distance between curves}
\usage{
maxDist(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
The maximum distance
}
\description{
This function computes the maximum distance between multiple curves
}
\keyword{internal}
