% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReactFixed.R
\name{parseBooleanCopy}
\alias{parseBooleanCopy}
\title{Copy of sybil:::.parseBoolean}
\usage{
parseBooleanCopy(gprRule, tokens = "()&|~")
}
\arguments{
\item{gprRule}{See \code{sybil:::.parseBoolean}.}

\item{token}{See \code{sybil:::.parseBoolean}.}
}
\value{
See \code{sybil:::.parseBoolean}.
}
\description{
This is the copy of sybil:::.parseBoolean, which is used to fix sybil::addReact
}
\keyword{internal}
