% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submnet.R
\name{rmGenes}
\alias{rmGenes}
\title{Generate a submodel by removing genes}
\usage{
rmGenes(model, genes)
}
\arguments{
\item{model}{An object of class \code{modelorg}.}

\item{genes}{A vector of genes to remove.}
}
\value{
The submodel.
}
\description{
This functions creates a submodel by removing genes from a given model. It is similar to \code{deleteModelGenes} from the COBRA Toolbox.
}
