% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submnet.R
\name{simulateSubmnet}
\alias{simulateSubmnet}
\title{Assess submodels built by gene removal in a condition}
\usage{
simulateSubmnet(sgd, mc.cores = 1, njt = NULL)
}
\arguments{
\item{sgd}{An object of class \code{scoreGeneDel}.}

\item{mc.cores}{The number of cores to use (at least 1), i.e. at most how many child processes will be run simultaneously. Default: 1.}

\item{njt}{An object of class \code{phylo} for colored plot of fitness weighting schema (see /tmp/njt*.RData files). Default: NULL.}
}
\value{
The significance levels
}
\description{
This function computes the significance levels of different rankings compared to the random ranking for gene removal in a condition.
}
\examples{
data(yarliSubmnets)
simulateSubmnet(yarliSubmnets$DN)
}
