% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yarli2GO}
\alias{yarli2GO}
\title{GO annotation retrieved from Uniprot}
\source{
Uniprot
}
\usage{
data(yarli2GO)
}
\description{
GO annotation retrieved from Uniprot
}
\references{
Tran, V.D.T. et al. (2018) Condition-specific series of metabolic sub-networks and its application for gene set enrichment analysis. Bioinformatics 35(13), 2258-2266.
}
