% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yarliGOdata}
\alias{yarliGOdata}
\title{topGOdata object built from yarli2GO}
\format{An object of class \code{"topGOdata"}.}
\source{
topGO_2.24.0, GO.db_3.2.2
}
\usage{
data(yarliGOdata)
}
\description{
topGOdata object built from yarli2GO
}
\references{
Tran, V.D.T. et al. (2018) Condition-specific series of metabolic  sub-networks and its application for gene set enrichment analysis. bioRxiv 200964.
}
