% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yarliSubmnets}
\alias{yarliSubmnets}
\title{Submodel construction example for Yarrowia lipolytica}
\usage{
data(yarliSubmnets)
}
\description{
Submodel construction example for Yarrowia lipolytica
}
\references{
Tran, V.D.T. et al. (2018) Condition-specific series of metabolic  sub-networks and its application for gene set enrichment analysis. Bioinformatics, bty929.
}
\author{
Van Du T. Tran
}
