% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{zscoreExpr}
\alias{zscoreExpr}
\title{Z-score of expression data}
\usage{
zscoreExpr(expr)
}
\arguments{
\item{expr}{A matrix of expression data (log2 counts), with gene per row and experiment per column.}
}
\value{
A matrix of z-score.
}
\description{
This function computes the z-score of expression data across different conditions.
}
\examples{
data(exprMaguire)
zExpr <- zscoreExpr(exprMaguire$expr)
}
